+++ packages/library-base/src/commonMain/kotlin/io/realm/notifications/ListChange.kt
@@ -17,11 +17,9 @@
 package io.realm.notifications
 
 import io.realm.RealmList
-import io.realm.RealmResults
 
 /**
- * This sealed interface describe the possible changes that can happen to a list collection,
- * currently [io.realm.RealmList] or [io.realm.RealmResults].
+ * This sealed interface describes the possible changes that can happen to a [RealmList] collection.
  *
  * The states are represented by the specific subclasses [InitialList], [UpdatedList] and
  * [DeletedList]. When the list is deleted an empty list is emitted instead of `null`.
@@ -30,19 +28,19 @@ import io.realm.RealmResults
  *
  * ```
  * // Variant 1: Switch on the sealed interface
- * realm.query<Person>().asFlow()
- *   .collect { it: ListChange<RealmResults<Person>> ->
+ * person.addresses.asFlow()
+ *   .collect { it: ListChange<Address> ->
  *       when(result) {
- *          is InitialList -> setUIList(it.list)
- *          is UpdatedList -> updateUIList(it) // Android RecyclerView knows how to animate ranges
- *          is DeletedList -> deleteUIList(it.list)
+ *          is InitialList -> setAddressesUIList(it.list)
+ *          is UpdatedList -> updateAddressesUIList(it) // Android RecyclerView knows how to animate ranges
+ *          is DeletedList -> deleteAddressesUIList(it.list)
  *       }
  *   }
  *
  *
  * // Variant 2: Just pass on the list
- * realm.query<Person>().asFlow()
- *   .collect { it: ListChange<RealmResults<Person>> ->
+ * person.addresses.asFlow()
+ *   .collect { it: ListChange<Address> ->
  *       handleChange(it.list)
  *   }
  * ```
@@ -53,8 +51,8 @@ import io.realm.RealmResults
  * to the appropriate subclass.
  *
  * ```
- * realm.query<Person>().asFlow()
- *   .collect { it: ListChange<RealmResults<Person>> ->
+ * person.addresses.asFlow()
+ *   .collect { it: ListChange<Address> ->
  *       when(result) {
  *          is InitialList -> setList(it.list)
  *          is UpdatedList -> { // Automatic cast to UpdatedList
@@ -70,77 +68,26 @@ import io.realm.RealmResults
  *   }
  * ```
  */
-public sealed interface ListChange<out T : List<*>> {
-    public data class Range(
-        /**
-         * The start index of this change range.
-         */
-        val startIndex: Int,
-        /**
-         * How many elements are inside this range.
-         */
-        val length: Int
-    )
-
-    public val list: T
+public sealed interface ListChange<T> {
+    public val list: RealmList<T>
 }
 
 /**
- * Initial event to be observed on a [RealmList] or [RealmResults] flow. It contains a reference to the
- * starting list state. Note, this state might be different than the list the flow was registered on,
- * if another thread or device updated the object in the meantime.
+ * Initial event to be observed on a [RealmList] flow. It contains a reference to the starting list
+ * state. Note, this state might be different than the list the flow was registered on, if another
+ * thread or device updated the object in the meantime.
  */
-public interface InitialList<T : List<*>> : ListChange<T>
+public interface InitialList<T> : ListChange<T>
 
 /**
- * [RealmList] or [RealmResults] flow event that describes that an update has been performed on to the
- * observed list. It provides a reference to the list and a set of properties that describes the changes
+ * [RealmList] flow event that describes that an update has been performed on the observed list. It
+ * provides a reference to the updated list and a set of properties that describes the changes
  * performed on the list.
  */
-public interface UpdatedList<T : List<*>> : ListChange<T> {
-    /**
-     * The deleted indices in the previous version of the collection. It will be set as a zero-sized
-     * array if no objects were deleted.
-     */
-    public val deletions: IntArray
-
-    /**
-     * The inserted indices in the new version of the collection. It will be set as a zero-sized
-     * array if no objects were inserted.
-     */
-    public val insertions: IntArray
-
-    /**
-     * The modified indices in the new version of the collection.
-     * <p>
-     * For [RealmResults], this means that one or more of the properties of the object at the given index were
-     * modified (or an object linked to by that object was modified). It will be set as a zero-sized
-     * array if no objects were changed.
-     */
-    public val changes: IntArray
-
-    /**
-     * The deleted ranges of objects in the previous version of the collection. It will be set as a zero-sized
-     * array if no objects were deleted.
-     */
-    public val deletionRanges: Array<ListChange.Range>
-
-    /**
-     * The inserted ranges of objects in the new version of the collection. It will be set as a zero-sized
-     * array if no objects were inserted.
-     */
-    public val insertionRanges: Array<ListChange.Range>
-
-    /**
-     * The modified ranges of objects in the new version of the collection. It will be set as a zero-sized
-     * array if no objects were changed.
-     */
-    public val changeRanges: Array<ListChange.Range>
-}
+public interface UpdatedList<T> : ListChange<T>, ListChangeSet
 
 /**
- * This interface describes the event is emitted deleted on a [RealmList] flow. The flow will terminate
- * after observing this event. This event would never be observed on a [RealmResults] as they cannot
- * be deleted.
+ * This event is emitted when the parent object owning the list has been deleted, which in turn also removes the list. The flow will terminate
+ * after observing this event.
  */
-public interface DeletedList<T : List<*>> : ListChange<T>
+public interface DeletedList<T> : ListChange<T>